Hoist shared board-import helpers into script/_board_import.py
